steps_Missing_Fix
We're missing a couple steps here, we have to connect to the server with the initial packet before we do anything, otherwise we're going to crash right away, or be stuck waiting. Its easier to know there is a server out there, before we do the work to convert an image and trying to send. Codes the same, just commenting so far.